New folder Structure added
