+++ src/App.jsx
@@ -1,0 +1,16 @@
+
+import Navbar from './components/Navbar';
+import Header from './components/features/Header';
+import Feature from './components/features/Feature';
+
+function App() {
+  return (
+    <div className="App">
+      <Navbar/>
+      <Header/>
+      <Feature/>
+    </div>
+  );
+}
+
+export default App;
