--- src/components/Button.js
@@ -1,13 +1,0 @@
-import React from 'react'
-
-const Button = () => {
-  return (
-    <div>
-        <button className='w-32  p-3 font-normal my-4 rounded-full bg-pink-500 text-white shadow-lg hover:shadow-pink-600 '>
-            Get started
-        </button>
-    </div>
-  )
-}
-
-export default Button;
