+++ src/components/nav/Contact.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Contact() {
+  return (
+    <div>Contact</div>
+  )
+}
+
+export default Contact
