+++ src/components/nav/Pricing.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Pricing() {
+  return (
+    <div>Pricing</div>
+  )
+}
+
+export default Pricing
