+++ src/components/nav/Products.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Products() {
+  return (
+    <div>Products</div>
+  )
+}
+
+export default Products
