+++ src/components/nav/Resources.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Resources() {
+  return (
+    <div>Resources</div>
+  )
+}
+
+export default Resources
