+++ src/components/nav/Solutions.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Solutions() {
+  return (
+    <div>Solutions</div>
+  )
+}
+
+export default Solutions
